Adds Config for token property on response object.
When using the Implicit Grant Type with OpenID Connect the JWT gets
return as `id_token`; the plug-in expects `access_token` to be the JWT.
See paulvanbladel/aurelia-auth#53.

This change is a work around in order to allow a user to set the property of the
response for use as the token if its something other than `access_token`.

This should not be a breaking change.